+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/Settings.kt
@@ -345,6 +345,11 @@ abstract class Settings {
      * Setting to control whether a DoH provider will be automatically selected when in Default Protection mode.
      */
     open var dohAutoselectEnabled: Boolean by UnsupportedSetting()
+
+    /**
+     * Comma-separated list of destination ports that the application should block connections to.
+     */
+    open var bannedPorts: String by UnsupportedSetting()
 }
 
 /**
@@ -414,6 +419,7 @@ data class DefaultSettings(
     override var certificateTransparencyMode: Int = 0,
     override var postQuantumKeyExchangeEnabled: Boolean? = null,
     override var dohAutoselectEnabled: Boolean = false,
+    override var bannedPorts: String = "",
 ) : Settings() {
     override val desktopModeEnabled: Boolean
         get() = getDesktopMode()