+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Core.kt
@@ -195,6 +195,7 @@ class Core(
             certificateTransparencyMode = FxNimbus.features.pki.value().certificateTransparencyMode,
             postQuantumKeyExchangeEnabled = FxNimbus.features.pqcrypto.value().postQuantumKeyExchangeEnabled,
             dohAutoselectEnabled = FxNimbus.features.doh.value().autoselectEnabled,
+            bannedPorts = FxNimbus.features.networkingBannedPorts.value().bannedPortList,
         )
 
         // Apply fingerprinting protection overrides if the feature is enabled in Nimbus